Ensure id set and sources vec stay in sync
Previously, an ID was always generated for each new URL by the source
manager, but there wouldn't always be a corresponding insertion into the
actual sources cache.

This commit fixes the issue by removing the added ID entry when an error
occurs and a source item could not be added to the cache.