remove checkboxes from issue templates
GitHub parses the checkboxes as a list of tasks required to resolve the
issue. It generates a progress bar on the issues page that doesn't mean
anything when these checkboxes are actually lists of tasks required to
post the issue.

To remove these, we leave the lists in place, but without checkboxes.